Update thirdparty/Q submodule (see note below)
NOTE: We are now committing to our own local branch of this. We should
then fork the repo and commit these changes to a branch in our fork,
so they're publicly available whenever this project is publicly
available. We haven't done that yet, but our commits to thirdparty/Q are
all very small, just to get the library building in our environment.