Add first example using Q lib to read a WAV, modify it, and write it back
This is just a test to make sure we can use the Q utilities to get and
work with the sample data from a WAV file. The only modification we
make to the WAV sample data is to reduce its amplitude to 10% of its
initial value, to make sure we can successfully make modifications.

Now that we have this we can work on applying some more fun techniques
in audio signal processing, and also learn more of the Q library
features.