Add simple program to play WAV file using ALSA API
This is mostly based on these sources:

Alessandro Ghedini's example (main source):
https://gist.github.com/ghedo/963382/815c98d1ba0eda1b486eb9d80d9a91a81d995283

Alex Via's blog post:
https://alexvia.com/post/003_alsa_playback/

ALSA PCM docs:
https://www.alsa-project.org/alsa-doc/alsa-lib/pcm.html

We've added some comments to explain the meaning of a few things.
It also has a little RAII class to mimic Go's defer command, that
we use for freeing a buffer.