Add code in kfr test program to re-write samples read from file
This demonstrates the kfr API for reading and writing files, which
allows us to read the samples from the channels of a wav (or other
format) file into an array of floating-point values, regardless of
the original sample format of the wav file. Then once we've done
whatever processing we want on the samples, we can write them back
out in whatever format we need.

The API is clean and user-friendly. This will be very useful, both
as a library to use and an example to learn from.