Add simple command line audio player
This loads a WAV file using kfr and plays the file's data using ALSA,
similarly to the simple_alsa_audio example. It takes an optional
command line argument specifying a path to an audio file, otherwise
it plays a sample file located in the media/ folder.

This will be a good starting point for other audio experiments,
since now we can process the file samples as floats and immediately
play them.

We will look at limiting the amount of data that is buffered,
in case we want to do some real-time processing. With the current
ALSA configuration, it appears that a large amount of data is
buffered before it gets played back.

Using kfr, we could also support other file formats.