Refactor: Create AlsaPlayer class that plays an AudioFile instance
This refactor is working towards organizing code for an application
with more complexity. Next we will play the audio on a separate thread,
to allow cancelling playback. Later we will add features like a
graphical display of playback intensity or a real-time spectral
analysis. And after that we'll aim towards filters we can adjust in
real time, starting with volume, bass and trebel, or something similar.