Play audio on separate thread and test early shutdown from main thread
This is just a simple test of using atomic shared variables to control
the playback thread state from the UI thread. It starts playing an ~30s
audio file on a separate thread, then stops the playback from the UI
thread.

TODO: Next will add a full-fledged (but simple) TUI using ncurses, to
allow the user to control the playback thread and also visualizing
information on the audio currently being played.