Move logging code outside of AlsaPlayer class due memory API bugs
It appears that calling any function that involves dynamic allocation
or that directly manipulates memory within the AlsaPlayer class
causes our call to the ALSA API to stop function correctly.

This is a bit of a mystery, but I've seen similar things before.
It looks like the problem only happens in the same *translation
unit* that contains the ALSA calls. We can create std::strings,
etc. in code outside that translation unit but in the same thread
and all is well. It may be that the ALSA include causes different
versions of the low-level memory functions (memcpy, malloc, etc.)
to be defined and/or linked, which conflict with the ones in the
C++ standard library.

It's not too hard to isolate the ALSA api calls to their own source
file for now. But we will also look for a proper solution to this.