Start implementing ncurses console interface for audio player
This is starting to build out some infrastructure for managing the
app and interface states. Currently it only implements loading a
fixed file, as a test. Next we will start fleshing out the rest of
the functionality, like playing back the file or requesting a file
name to try to load.