Add clang sanitizers; debuging; add alsa official example; update kfr
Note: Addressing the comment in audio_player_main.cpp, whenever we
get the filename from user input using ncurses, the sound level
meter stops working, but the sound file still plays back. This looks
to me like either undefined behavior or some global state that is
being manipulated by the ncurses calls. Note that the clang UB and
address sanitizers don't find any problems in code that we have
compiled, when this is happening. So I've hardcoded the test file
path for now, and will look into this issue further.